cpuid additions and minor tweaks

+++ include/sled/enum.h
@@ -167,7 +167,7 @@ struct enum_struct {
         return os;
       }
     }
-    os << "Invalid<" << e.v << ">";
+    os << "Invalid<" << (int)e.v << ">";
     return os;
   }
